pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bq-features-demo"
version = "0.1.0"
description = "CLI to deploy BigQuery Studio demo infrastructure and finance-oriented demo material"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Lumax" }]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-storage>=2.0",
    "google-cloud-bigquery-datatransfer>=3.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
bqdemo = "bq_features.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bq_features = ["demo_data/*.csv"]

[tool.ruff]
line-length = 100
target-version = "py310"
